requirements.txt
langchain-google-genai==2.0.1
langchain-community==0.3.1
langchain-core==0.3.10
langgraph==0.2.39
faiss-cpu==1.8.0
tree-sitter==0.23.0
tree-sitter-python==0.23.0
mcp==1.0.0
pydantic>=2.0.0
python-dotenv
